+++ tests/bootstrap.php
@@ -13,4 +13,4 @@ declare(strict_types=1);
 
 require __DIR__ . '/../vendor/autoload.php';
 
-\date_default_timezone_set('UTC');
+\date_default_timezone_set('Europe/Berlin');